refactor: 패키지 위치 수정
SL-115
