+++ backend/src/main/java/com/devdive/backend/study/application/service/StudyService.java
@@ -2,7 +2,7 @@ package com.devdive.backend.study.application.service;
 
 import com.devdive.backend.study.application.dto.StudyCreateDto;
 import com.devdive.backend.study.application.port.in.StudyUseCase;
-import com.devdive.backend.study.application.port.out.persistence.LoadStudyPort;
+import com.devdive.backend.study.application.port.out.LoadStudyPort;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
